🎨 budget Float to Double / SetBudgetOrDistanceDialog text bug (when big number) fix

+++ core/data/firebase-firestore/src/main/java/com/newpaper/somewhere/core/firebase_firestore/model/SpotFirestore.kt
@@ -27,7 +27,7 @@ internal data class SpotFirestore(
     val startTime: String? = null,
     val endTime: String? = null,
 
-    val budget: Float = 0.0f,
+    val budget: Double = 0.0,
     val travelDistance: Float = 0.0f,
     val memoText: String? = null
 ){
